refactor(agent): extract standalone helpers to dkg-agent-utils.ts
dkg-agent.ts was the largest source file in packages/agent (~7,240
lines). This mechanical extraction moves a set of pure standalone
helpers — N-Quads parsing/serialization, JSON-LD to RDF conversion,
and the Merkle-root sync verifier — into a sibling
`dkg-agent-utils.ts` module.

Scope (pure code-motion, no behavior change):

- New: packages/agent/src/dkg-agent-utils.ts (~395 lines) containing
  splitNQuadLine, strip, stripLiteral, unescapeLiteralContent,
  parseNQuads, getJsonld, assignUrisToBlankNodes, jsonLdToQuads,
  verifySyncedData + the JsonLdDocument/JsonLdContent type aliases
  and the DKG_NS/RDF_TYPE constants.

- dkg-agent.ts imports the symbols it still uses (strip, stripLiteral,
  jsonLdToQuads, JsonLdContent) from ./dkg-agent-utils.js; the rest
  are only referenced transitively by the extracted module.

Net: dkg-agent.ts drops by ~370 lines (377 deletions, 4 insertions).
The N-Quads parser is intentionally duplicated with the existing copy
in sync-verify-worker-impl.ts — that duplication pre-dates this
extraction and is out of scope for this PR.

Verified via `tsc --noEmit` on packages/agent — clean. The targeted
sync-verify-worker warnings and Hardhat-harness failures in the full
test suite are pre-existing on main and unrelated to this change.

This sets up dkg-agent.ts for further incremental decomposition
(e.g. peeling off the publish/sync pipelines into their own modules)
without the noise of today's mechanical move.

Made-with: Cursor
